+++ src/main/java/com/webeye/backend/review/application/ReviewService.java
@@ -3,8 +3,8 @@ package com.webeye.backend.review.application;
 import com.webeye.backend.review.dto.request.ReviewSummaryRequest;
 import com.webeye.backend.review.dto.response.ReviewResponse;
 import com.webeye.backend.review.infrastructure.clovaX.ClovaXClient;
-import com.webeye.backend.review.infrastructure.clovaX.domain.ContentType;
-import com.webeye.backend.review.infrastructure.clovaX.domain.Role;
+import com.webeye.backend.review.infrastructure.clovaX.model.ContentType;
+import com.webeye.backend.review.infrastructure.clovaX.model.Role;
 import com.webeye.backend.review.infrastructure.clovaX.dto.request.ClovaXContent;
 import com.webeye.backend.review.infrastructure.clovaX.dto.request.ClovaXMessage;
 import com.webeye.backend.review.infrastructure.clovaX.dto.request.ClovaXRequest;
