+++ src/main/java/com/webeye/backend/review/infrastructure/clovaX/dto/request/ClovaXMessage.java
@@ -1,6 +1,6 @@
 package com.webeye.backend.review.infrastructure.clovaX.dto.request;
 
-import com.webeye.backend.review.infrastructure.clovaX.domain.Role;
+import com.webeye.backend.review.infrastructure.clovaX.model.Role;
 
 import java.util.List;
